add ScriptFileContext.cc
